+++ autoblocks/_impl/tracer/decorators.py
@@ -1,10 +1,8 @@
 import asyncio
 import functools
-import traceback
 from typing import Any
 from typing import Callable
 
-import orjson
 from opentelemetry import trace
 from opentelemetry.baggage import set_baggage
 from opentelemetry.context import attach
@@ -14,31 +12,7 @@ from opentelemetry.context import get_current
 from autoblocks._impl.context_vars import test_case_run_context_var
 from autoblocks._impl.tracer.util import SpanAttribute
 from autoblocks._impl.util import cuid_generator
-
-
-def orjson_default(o: Any) -> Any:
-    if hasattr(o, "model_dump_json") and callable(o.model_dump_json):
-        # pydantic v2
-        return orjson.loads(o.model_dump_json())
-    elif hasattr(o, "json") and callable(o.json):
-        # pydantic v1
-        return orjson.loads(o.json())
-    elif isinstance(o, Exception):
-        return "".join(
-            traceback.format_exception(
-                type(o),
-                o,
-                o.__traceback__,
-            )
-        )
-    raise TypeError
-
-
-def serialize(value: Any) -> str:
-    try:
-        return orjson.dumps(value, default=orjson_default).decode("utf-8")
-    except Exception:
-        return "\\{\\}"
+from autoblocks._impl.util import serialize
 
 
 def trace_app(app_slug: str, environment: str) -> Callable[[Callable[..., Any]], Callable[..., Any]]:
